==> 01 Boliga Scrape - Rådata.ipynb <==
# Scrape of Boliga

import requests
import pandas as pd
import tqdm
import json

url = 'https://api.boliga.dk/api/v2/search/results?pageSize=50&sort=zipCode-a&page=1'
response = requests.get(url, 'html.parser').json()

#Calculate nr of iterations
totalpages = response['meta']['totalPages']

response['results'][0]

# TEST! 
import time
import tqdm
import scraping_class

logfile = 'logfile_boliga.csv'
connector = scraping_class.Connector(logfile)

df_old = pd.DataFrame()

#Scrape the page:
for page in tqdm.tqdm(range(1, 50)):
    time.sleep(0.5)
    url = 'https://api.boliga.dk/api/v2/search/results?pageSize=50&sort=zipCode-a&page={p}'.format(p = page)
    response,call_id = connector.get(url,'Boligdata')
    response = response.json()
    response = response['results']
    df_new = pd.DataFrame(response)
    df_old = pd.concat([df_old, df_new], axis = 0)

### Scrape 

import time
import tqdm
import scraping_class

logfile = 'logfile_boliga.csv'
connector = scraping_class.Connector(logfile)

df_old = pd.DataFrame()

#Scrape the page:
for page in tqdm.tqdm(range(1, totalpages)):
    time.sleep(0.5)
    url = 'https://api.boliga.dk/api/v2/search/results?pageSize=50&sort=zipCode-a&page={p}'.format(p = page)
    response,call_id = connector.get(url,'Boligdata')
    response = response.json()
    response = response['results']
    df_new = pd.DataFrame(response)
    df_old = pd.concat([df_old, df_new], axis = 0)

#Join coordinates
df['coordinate'] = df['latitude'].astype(str) + ', ' + df['longitude'].astype(str)

#Save data
df.to_csv('Boliga Scraped - Rådata.csv', index = False)
