==> 03 Boliga - Behandling 2 - Befolkning, Ledig, Udd, Indkomst.ipynb <==
# Behandling 2

import pandas as pd

pd.options.display.max_columns = 100

df = pd.read_csv('Boliga - Behandling 1 - Kom, Reg, Type.csv')

df_befolkning = pd.read_csv('DST_Excel/Befolkning pr kommune.csv', sep = ';', encoding='latin-1', skiprows=1)
COL = ['Kommune', 'Befolkning']
df_befolkning.columns = COL

#Merge befolkningsdata onto master
df = df.merge(df_befolkning, how='left', left_on=['Kommune'],
                           right_on=['Kommune'], validate = 'm:1')

df = df[df.Befolkning.isnull() != True]


#Load Relative unemplyment pr. capita data
df_led = pd.read_excel('DST_Excel/Relativ Ledighed og Indbyggertal 30-59år .xlsx',  sep = ';', encoding='latin-1')

#Merge unemploymentdata onto master
df = df.merge(df_led, how='left', left_on=['Kommune'],
                           right_on=['Kommuner'], validate = 'm:1')

df = df.drop(columns = ['Kommuner', 'Indbygger'])
df = df[df.Ledige.isnull() != True]

#Load Relative education pr. capita data
df_udd = pd.read_excel('DST_Excel/Relativ Uddannelse og Kommune 30-69år.xlsx',  sep = ';', encoding='latin-1')

#Merge education data onto master 
df = df.merge(df_udd, how='left', left_on=['Kommune'],
                           right_on=['Kommuner'], validate = 'm:1')

df = df.drop(columns = ['Kommuner'])
df = df[df.Grundskole.isnull() != True]


df_in = pd.read_excel('DST_Excel/Indkomstdata.xlsx')

df = df.merge(df_in, how='left', left_on=['Kommune'],
                           right_on=['Kommune'], validate = 'm:1')



df.to_csv('Boliga - Behandling 2, Ledig, Udd, Indbygger, indkomst.csv', index = False)
