==> 07 Final_boliga_data_allvariables.ipynb <==
import pandas as pd
df = pd.read_csv('howlongdf.csv')
df2 = pd.read_csv('Boliga - Behandling 4 - Med etnosociodata.csv')

pd.options.display.max_columns = 100

df = df[['id','hvorlangt_total']].copy()

df.head()

df.shape

df = df.drop_duplicates()

df.shape

df2 = df2.drop_duplicates()

df2.shape

finaldf = df2.merge(df, how='left', left_on = df2['id'], right_on = df['id'], validate= 'm:1')

df2['id'].unique()

finaldf.shape

finaldf = finaldf.drop(columns=['id_y']).copy()

finaldf.head()

from ast import literal_eval

poitypes = ["lake", "forest", "doctor", "supermarket", "school", "daycare", "hospital", "metro",
            "train", "stop", "strain", "junction", "pharmacy", "library", "coast", "airport"]
def get_lakekey(key_column):
    key_column = literal_eval(key_column)
    return key_column['lake']
def get_forestkey(key_column):
    key_column = literal_eval(key_column)
    return key_column['forest']
def get_doctorkey(key_column):
    key_column = literal_eval(key_column)
    return key_column['doctor']
def get_supermarketkey(key_column):
    key_column = literal_eval(key_column)
    return key_column['supermarket']
def get_schoolkey(key_column):
    key_column = literal_eval(key_column)
    return key_column['school']
def get_daycarekey(key_column):
    key_column = literal_eval(key_column)
    return key_column['daycare']
def get_hospitalkey(key_column):
    key_column = literal_eval(key_column)
    return key_column['hospital']
def get_metrokey(key_column):
    key_column = literal_eval(key_column)
    return key_column['metro']
def get_trainkey(key_column):
    key_column = literal_eval(key_column)
    return key_column['train']
def get_stopkey(key_column):
    key_column = literal_eval(key_column)
    return key_column['stop']
def get_strainkey(key_column):
    key_column = literal_eval(key_column)
    return key_column['strain']
def get_junctionkey(key_column):
    key_column = literal_eval(key_column)
    return key_column['junction']
def get_pharmacykey(key_column):
    key_column = literal_eval(key_column)
    return key_column['pharmacy']
def get_librarykey(key_column):
    key_column = literal_eval(key_column)
    return key_column['library']
def get_coastkey(key_column):
    key_column = literal_eval(key_column)
    return key_column['coast']
def get_airportkey(key_column):
    key_column = literal_eval(key_column)
    return key_column['airport']


finaldf['lake_distance'] = finaldf['hvorlangt_total'].apply(get_lakekey)

poitypes = ["lake", "forest", "doctor", "supermarket", "school", "daycare", "hospital", "metro",
            "train", "stop", "strain", "junction", "pharmacy", "library", "coast", "airport"]
finaldf['forest_distance'] = finaldf['hvorlangt_total'].apply(get_forestkey)
finaldf['doctor_distance'] = finaldf['hvorlangt_total'].apply(get_doctorkey)
finaldf['supermarket_distance'] = finaldf['hvorlangt_total'].apply(get_supermarketkey)
finaldf['school_distance'] = finaldf['hvorlangt_total'].apply(get_schoolkey)

finaldf['daycare_distance'] = finaldf['hvorlangt_total'].apply(get_daycarekey)
finaldf['hospital_distance'] = finaldf['hvorlangt_total'].apply(get_hospitalkey)
finaldf['metro_distance'] = finaldf['hvorlangt_total'].apply(get_metrokey)
finaldf['train_distance'] = finaldf['hvorlangt_total'].apply(get_trainkey)

finaldf['pharmacy_distance'] = finaldf['hvorlangt_total'].apply(get_pharmacykey)
finaldf['library_distance'] = finaldf['hvorlangt_total'].apply(get_librarykey)
finaldf['coast_distance'] = finaldf['hvorlangt_total'].apply(get_coastkey)
finaldf['airport_distance'] = finaldf['hvorlangt_total'].apply(get_airportkey)

#finaldf['stop_distance'] = finaldf['hvorlangt_total'].apply(get_stopkey)
finaldf['strain_distance'] = finaldf['hvorlangt_total'].apply(get_strainkey)
finaldf['junction_distance'] = finaldf['hvorlangt_total'].apply(get_junctionkey)

finaldf.head()

finaldf.shape

finaldf = finaldf.drop(columns=['hvorlangt_total']).copy()

for i in finaldf.isna().sum():
    print(i)

finaldf.to_csv('Final_Boligdata.csv')
