==> 08 Sidste behandling - oprydning i features.ipynb <==
# Sidste behandling - oprydning i features

import pandas as pd

df = pd.read_csv('Final_Boligdata.csv')

pd.options.display.max_columns = None


### Delete observations with long 'daysForSale'

#Set limit to 1095 days / 3 years
df = df[df.daysForSale < 1095]

# Convert ethinicity of population to relative measures. 
df['Total_reported'] = df['Total_reported'] / df['Befolkning'] * 100

#Convert to km instead of m:
df['lake_distance'] = df['lake_distance'] / 1000
df['forest_distance'] = df['forest_distance'] / 1000
df['doctor_distance'] = df['doctor_distance'] / 1000
df['supermarket_distance'] = df['supermarket_distance'] / 1000
df['school_distance'] = df['school_distance'] / 1000
df['daycare_distance'] = df['daycare_distance'] / 1000
df['hospital_distance'] = df['hospital_distance'] / 1000
df['pharmacy_distance'] = df['pharmacy_distance'] / 1000
df['library_distance'] = df['library_distance'] / 1000
df['coast_distance'] = df['coast_distance'] / 1000
df['junction_distance'] = df['junction_distance'] / 1000
df['train_distance'] = df['train_distance'] / 1000



pd.options.display.max_columns = None
pd.options.display.max_rows = 200

### Final cleanup in columns

df_final = df.drop( columns = ['Unnamed: 0', 'key_0', 'city', 'createdDate', 'floor', 'id_x',
                              'daysForSale', 'latitude', 'longitude', 'municipality',
                              'street', 'zipCode', 'coordinate', 'Region', 'Befolkning',
                              'danish_origin', 'immigrants_Western', 'immigrants_Non_Western',
                              'descendants_Western', 'descendants_Non_Western', 'Murder&Arson_reported',
                              'Murder&Arson_charged', 'Breaking&entering_reported', 'Breaking&entering_charged',
                              'Narcotics_reported', 'Narcotics_charged', 'Robbery_reported', 'Robbery_charged',
                              'Indecency_reported', 'Indecency_charged', 'Theft_reported', 'Theft_charged',
                              'Assault_reported', 'Assault_charged', 'Total_charged', 'Population_density(pop/km2)',
                              'expenditure_per_capita', 'Service_expenses_per_capita', 'church_tax', 'land_tax(‰)',
                              'land_value_per_capita', 'taxable_income_per_capita', 'expenses_development',
                              'expenses_transport&infrastructure', 'expenses_public_school_per_capita',
                              'expenses_daycare_per_capita', 'long_term_debt_per_capita', 'liquid_assets_per_capita',
                              'expenses_daycare_per_0-13yo(brutto)', 'expenses_daycare_per_0-13yo(netto)', 'monthly_price_daycare(3-5yo)',
                              'education_expenses(6-16yo)', 'number_of_public_schools', 'number_of_ordinary_classes',
                              'library_expenses_per_capita','expenses_elder(+67yo)', 'sickness_benefits_per_17-64yo',
                               'expenses_for_assistance(17-64yo)', 'expenses_for_reschooling(17-64yo)', 'metro_distance',
                               'strain_distance', 'airport_distance'])

df_final.shape

df_final.to_csv('Boliga - Final for training.csv', index = False)
